requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/ # https://github.com/microsoft/onnxruntime/issues/21684
conformer==0.3.2
deepspeed==0.15.1; sys_platform == 'linux'
diffusers==0.29.0
fastapi==0.110.0
face-alignment==1.4.1
facexlib==0.3.0
fastapi-cli==0.0.4
gdown==5.1.0
gfpgan==1.3.8
gradio==4.31.5
grpcio==1.57.0
grpcio-tools==1.57.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
inflect==7.3.1
imageio==2.34.1
imageio-ffmpeg==0.5.1
kornia==0.7.2
librosa==0.10.2
lightning==2.2.4
matplotlib==3.7.5
modelscope==1.20.0
networkx==3.1
omegaconf==2.3.0
onnx==1.16.0
onnxruntime-gpu==1.18.0; sys_platform == 'linux'
onnxruntime==1.18.0; sys_platform == 'darwin' or sys_platform == 'win32'
opencv-python-headless==4.10.0.84
openai-whisper==20231117
protobuf==4.25
pyarrow==18.1.0
pydantic==2.7.4
pyworld==0.3.4
rich==13.7.1
soundfile==0.12.1
tensorboard==2.14.0
tensorrt-cu12==10.0.1; sys_platform == 'linux'
tensorrt-cu12-bindings==10.0.1; sys_platform == 'linux'
tensorrt-cu12-libs==10.0.1; sys_platform == 'linux'
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
transformers==4.40.1
uvicorn==0.30.0
wetext==0.0.4
wget==3.2
yacs==0.1.8
langchain==0.3.27
langchain-core==0.3.72
langchain-community==0.3.27
langchain-deepseek==0.1.4
python-dotenv
